Chore: refactor workflows
